+++ jquery.handsontable.js
@@ -6,7 +6,7 @@
  * Licensed under the MIT license.
  * http://handsontable.com/
  *
- * Date: Tue Mar 26 2013 11:41:47 GMT+0100 (Central European Standard Time)
+ * Date: Tue Mar 26 2013 12:27:34 GMT+0100 (Central European Standard Time)
  */
 /*jslint white: true, browser: true, plusplus: true, indent: 4, maxerr: 50 */
 
@@ -2838,17 +2838,24 @@ Handsontable.SelectionPoint.prototype.arr = function (arr) {
 /**
  * Default text renderer
  * @param {Object} instance Handsontable instance
- * @param {Element} td Table cell where to render
+ * @param {Element} TD Table cell where to render
  * @param {Number} row
  * @param {Number} col
  * @param {String|Number} prop Row object property name
  * @param value Value to render (remember to escape unsafe HTML before inserting to DOM!)
  * @param {Object} cellProperties Cell properites (shared by cell renderer and editor)
  */
-Handsontable.TextRenderer = function (instance, td, row, col, prop, value, cellProperties) {
+Handsontable.TextRenderer = function (instance, TD, row, col, prop, value, cellProperties) {
   var escaped = Handsontable.helper.stringify(value);
-  escaped = escaped.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;").replace(/"/g, "&quot;").replace(/'/g, "&#039;"); //escape html special chars
-  td.innerHTML = escaped.replace(/\n/g, '<br/>');
+  if (escaped.match(/\n/)) {
+    escaped = escaped.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;").replace(/"/g, "&quot;").replace(/'/g, "&#039;"); //escape html special chars
+    TD.innerHTML = escaped.replace(/\n/g, '<br/>');
+  }
+  else {
+    Handsontable.helper.empty(TD); //TODO identify under what circumstances this line can be removed
+    TD.appendChild(document.createTextNode(escaped));
+    //this is faster than innerHTML. See: https://github.com/warpech/jquery-handsontable/wiki/JavaScript-&-DOM-performance-tips
+  }
 };
 /**
  * Autocomplete renderer
@@ -2866,7 +2873,8 @@ Handsontable.AutocompleteRenderer = function (instance, TD, row, col, prop, valu
 
   var ARROW = document.createElement('DIV');
   ARROW.className = 'htAutocompleteArrow';
-  ARROW.innerHTML = '&#x25BC;';
+  ARROW.appendChild(document.createTextNode('\u25BC'));
+  //this is faster than innerHTML. See: https://github.com/warpech/jquery-handsontable/wiki/JavaScript-&-DOM-performance-tips
 
   if (!instance.acArrowListener) {
     //not very elegant but easy and fast
@@ -2880,7 +2888,8 @@ Handsontable.AutocompleteRenderer = function (instance, TD, row, col, prop, valu
 
   if (!TEXT.firstChild) { //http://jsperf.com/empty-node-if-needed
     //otherwise empty fields appear borderless in demo/renderers.html (IE)
-    TEXT.innerHTML = '&nbsp;';
+    TEXT.appendChild(document.createTextNode('\u00A0')); //\u00A0 equals &nbsp; for a text node
+    //this is faster than innerHTML. See: https://github.com/warpech/jquery-handsontable/wiki/JavaScript-&-DOM-performance-tips
   }
 
   TEXT.appendChild(ARROW);
@@ -2924,7 +2933,8 @@ Handsontable.CheckboxRenderer = function (instance, TD, row, col, prop, value, c
     TD.appendChild(INPUT);
   }
   else {
-    TD.innerHTML = "#bad value#";
+    TD.appendChild(document.createTextNode('#bad value#'));
+    //this is faster than innerHTML. See: https://github.com/warpech/jquery-handsontable/wiki/JavaScript-&-DOM-performance-tips
   }
 
   if (!instance.checkboxInputMousedownListener) {
@@ -2951,23 +2961,25 @@ Handsontable.CheckboxRenderer = function (instance, TD, row, col, prop, value, c
 /**
  * Numeric cell renderer
  * @param {Object} instance Handsontable instance
- * @param {Element} td Table cell where to render
+ * @param {Element} TD Table cell where to render
  * @param {Number} row
  * @param {Number} col
  * @param {String|Number} prop Row object property name
  * @param value Value to render (remember to escape unsafe HTML before inserting to DOM!)
  * @param {Object} cellProperties Cell properites (shared by cell renderer and editor)
  */
-Handsontable.NumericRenderer = function (instance, td, row, col, prop, value, cellProperties) {
+Handsontable.NumericRenderer = function (instance, TD, row, col, prop, value, cellProperties) {
   if (typeof value === 'number') {
     if (typeof cellProperties.language !== 'undefined') {
       numeral.language(cellProperties.language)
     }
-    td.innerHTML = numeral(value).format(cellProperties.format || '0'); //docs: http://numeraljs.com/
-    td.className = 'htNumeric';
+    Handsontable.helper.empty(TD); //TODO identify under what circumstances this line can be removed
+    TD.className = 'htNumeric';
+    TD.appendChild(document.createTextNode(numeral(value).format(cellProperties.format || '0'))); //docs: http://numeraljs.com/
+    //this is faster than innerHTML. See: https://github.com/warpech/jquery-handsontable/wiki/JavaScript-&-DOM-performance-tips
   }
   else {
-    Handsontable.TextRenderer(instance, td, row, col, prop, value, cellProperties);
+    Handsontable.TextRenderer(instance, TD, row, col, prop, value, cellProperties);
   }
 };
 function HandsontableTextEditorClass(instance) {