+++ src/renderers/checkboxRenderer.js
@@ -35,7 +35,8 @@ Handsontable.CheckboxRenderer = function (instance, TD, row, col, prop, value, c
     TD.appendChild(INPUT);
   }
   else {
-    TD.innerHTML = "#bad value#";
+    TD.appendChild(document.createTextNode('#bad value#'));
+    //this is faster than innerHTML. See: https://github.com/warpech/jquery-handsontable/wiki/JavaScript-&-DOM-performance-tips
   }
 
   if (!instance.checkboxInputMousedownListener) {